+++ src/Client.php
@@ -53,6 +53,10 @@ class Client implements ClientInterface, \Psr\Http\Client\ClientInterface
      *   into relative URIs. Can be a string or instance of UriInterface.
      * - transport_sharing: (string|null) Transport sharing mode for the
      *   default handler. Accepts TransportSharing::* or null. Defaults to null.
+     * - max_host_connections: (int|null) Maximum concurrent connections per
+     *   host for the default cURL multi handler.
+     * - max_total_connections: (int|null) Maximum concurrent connections for
+     *   the default cURL multi handler.
      * - **: any request option
      *
      * @param array $config Client configuration settings.
@@ -61,16 +65,33 @@ class Client implements ClientInterface, \Psr\Http\Client\ClientInterface
      */
     public function __construct(array $config = [])
     {
+        $handlerOptions = [];
+        foreach (['max_host_connections', 'max_total_connections'] as $capOption) {
+            if (\array_key_exists($capOption, $config)) {
+                if ($config[$capOption] !== null) {
+                    $handlerOptions[$capOption] = $config[$capOption];
+                }
+
+                unset($config[$capOption]);
+            }
+        }
+
         $transportSharing = \array_key_exists('transport_sharing', $config) ? $config['transport_sharing'] : null;
         $transportSharingMode = CurlShareHandleState::normalizeMode($transportSharing, 'transport_sharing');
         unset($config['transport_sharing']);
 
         if (!isset($config['handler'])) {
-            $config['handler'] = $transportSharingMode === TransportSharing::NONE
+            if ($transportSharingMode !== TransportSharing::NONE) {
+                $handlerOptions['transport_sharing'] = $transportSharingMode;
+            }
+
+            $config['handler'] = $handlerOptions === []
                 ? HandlerStack::create()
-                : HandlerStack::create(Utils::chooseHandler(['transport_sharing' => $transportSharingMode]));
+                : HandlerStack::create(Utils::chooseHandler($handlerOptions));
         } elseif (!\is_callable($config['handler'])) {
             throw new InvalidArgumentException('handler must be a callable');
+        } elseif ($handlerOptions !== []) {
+            throw new InvalidArgumentException('The "max_host_connections" and "max_total_connections" client options require Guzzle to create the default handler. Configure the option on CurlMultiHandler when providing a custom handler.');
         } elseif ($transportSharingMode === TransportSharing::HANDLER_REQUIRE) {
             throw new InvalidArgumentException('The "transport_sharing" client option can only require sharing when Guzzle creates the default handler. Configure the "transport_sharing" option on CurlHandler or CurlMultiHandler when providing a custom cURL handler.');
         }